+++ spec/vectors/k3_secret-pw_spec.rb
@@ -53,7 +53,7 @@ RSpec.describe "PASERK k3.secret-pw Test Vectors" do
       Paseto::Paserk.from_paserk(paserk: paserk, password: password)
     end.to raise_error(Paseto::InvalidAuthenticator)
 
-    if Paseto.rbnacl?
+    if Paseto::HAS_RBNACL
       expect do
         pbkw.encode(Paseto::V4::Local.new(ikm: 0.chr * 32), options)
       end.to raise_error(Paseto::LucidityError)
@@ -70,7 +70,7 @@ RSpec.describe "PASERK k3.secret-pw Test Vectors" do
       Paseto::Paserk.from_paserk(paserk: paserk, password: password)
     end.to raise_error(Paseto::InvalidAuthenticator)
 
-    if Paseto.rbnacl?
+    if Paseto::HAS_RBNACL
       expect do
         pbkw.encode(Paseto::V4::Local.new(ikm: 0.chr * 32), options)
       end.to raise_error(Paseto::LucidityError)
@@ -85,7 +85,7 @@ RSpec.describe "PASERK k3.secret-pw Test Vectors" do
 
     expect { pbkw.decode(paserk) }.to raise_error(Paseto::LucidityError)
 
-    if Paseto.rbnacl?
+    if Paseto::HAS_RBNACL
       expect do
         pbkw.encode(Paseto::V4::Local.new(ikm: 0.chr * 32), options)
       end.to raise_error(Paseto::LucidityError)