+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove support for OpenSSL 1.1.1
 - Remove support for ruby/openssl 3.0.x
 - Refactor how version protocols are implemented to greatly improve sorbet coverage
+- `Paseto.rbnacl?` is replaced by `Paseto::HAS_RBNACL`
 
 ## [0.1.2]